+++ src/main/scala/is/hail/expr/ir/functions/MathFunctions.scala
@@ -24,12 +24,17 @@ object MathFunctions extends RegistryFunctions {
 
   def mod(x: Float, y: Float): Float = {
     val t = x % y
-    if (x >= 0 && y > 0 || x <= 0 && y < 0 || t == 0) t else t + y
+    if (t < 0) t + y else t
   }
   def mod(x: Double, y: Double): Double = {
     val t = x % y
-    if (x >= 0 && y > 0 || x <= 0 && y < 0 || t == 0) t else t + y
+    if (t < 0) t + y else t
   }
+
+  def pow(x: Int, y: Int): Double = math.pow(x, y)
+  def pow(x: Long, y: Long): Double = math.pow(x, y)
+  def pow(x: Float, y: Float): Double = math.pow(x, y)
+  def pow(x: Double, y: Double): Double = math.pow(x, y)
 
   def floorDiv(x: Int, y: Int): Int = java.lang.Math.floorDiv(x, y)
 
@@ -56,17 +61,11 @@ object MathFunctions extends RegistryFunctions {
     registerIR("toInt64", tnum("T"))(x => Cast(x, TInt64()))
     registerIR("toFloat32", tnum("T"))(x => Cast(x, TFloat32()))
     registerIR("toFloat64", tnum("T"))(x => Cast(x, TFloat64()))
-    registerCode("toInt32", TBoolean(), TInt32()){ case x: Code[Boolean] => x.toI }
-    registerCode("toInt64", TBoolean(), TInt32()){ case x: Code[Boolean] => x.toI.toL }
-    registerCode("toFloat32", TBoolean(), TInt32()){ case x: Code[Boolean] => x.toI.toF }
-    registerCode("toFloat64", TBoolean(), TInt32()){ case x: Code[Boolean] => x.toI.toD }
 
-    registerCode("+", tnum("T"), tnum("T")){
-      case x: Code[Int] => (x < 0).mux(-x, x)
-      case x: Code[Long] => (x < 0L).mux(-x, x)
-      case x: Code[Float] => (x < 0.0f).mux(-x, x)
-      case x: Code[Double] => (x < 0.0).mux(-x, x)
-    }
+    registerScalaFunction("**", TInt32(), TInt32(), TFloat64())(thisClass, "pow")
+    registerScalaFunction("**", TInt64(), TInt64(), TFloat64())(thisClass, "pow")
+    registerScalaFunction("**", TFloat32(), TFloat32(), TFloat64())(thisClass, "pow")
+    registerScalaFunction("**", TFloat64(), TFloat64(), TFloat64())(thisClass, "pow")
 
     registerScalaFunction("exp", TFloat64(), TFloat64())(mathPackageClass, "exp")
     registerScalaFunction("log10", TFloat64(), TFloat64())(mathPackageClass, "log10")
@@ -74,7 +73,6 @@ object MathFunctions extends RegistryFunctions {
     registerScalaFunction("log", TFloat64(), TFloat64())(mathPackageClass, "log")
     registerScalaFunction("log", TFloat64(), TFloat64(), TFloat64())(thisClass, "log")
     registerScalaFunction("pow", TFloat64(), TFloat64(), TFloat64())(mathPackageClass, "pow")
-    registerScalaFunction("**", TFloat64(), TFloat64(), TFloat64())(mathPackageClass, "pow")
     registerScalaFunction("gamma", TFloat64(), TFloat64())(thisClass, "gamma")
 
     registerScalaFunction("binomTest", TInt32(), TInt32(), TFloat64(), TInt32(), TFloat64())(statsPackageClass, "binomTest")
@@ -111,9 +109,9 @@ object MathFunctions extends RegistryFunctions {
     registerScalaFunction("ceil", TFloat64(), TFloat64())(thisClass, "ceil")
 
     registerJavaStaticFunction("%", TInt32(), TInt32(), TInt32())(jMathClass, "floorMod")
-    registerJavaStaticFunction("%", TInt64(), TInt32(), TInt64())(jMathClass, "floorMod")
-    registerJavaStaticFunction("%", TFloat32(), TInt32(), TFloat32())(jMathClass, "floorMod")
-    registerJavaStaticFunction("%", TFloat64(), TInt32(), TFloat64())(jMathClass, "floorMod")
+    registerJavaStaticFunction("%", TInt64(), TInt64(), TInt64())(jMathClass, "floorMod")
+    registerJavaStaticFunction("%", TFloat32(), TFloat32(), TFloat32())(thisClass, "mod")
+    registerJavaStaticFunction("%", TFloat64(), TFloat64(), TFloat64())(thisClass, "mod")
 
     registerScalaFunction("isnan", TFloat64(), TBoolean())(thisClass, "isnan")
